pyproject.toml
[project]
name = "chinese-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.122.0",
    "jieba>=0.42.1",
    "psycopg2-binary>=2.9.0",
    "openai>=2.8.1",
    "openai-agents>=0.0.19",
    "python-multipart>=0.0.20",
    "uvicorn>=0.38.0",
]
